+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Breakpoint Sync
 
 - Add command to remove breakpoints from outside the workspace
+- Add command to remove all breakpoints from current file
 
 ## v0.0.2 - 2020-06-30